% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_h2o_confusion_metrics.R
\name{custom_h2o_confusion_metrics}
\alias{custom_h2o_confusion_metrics}
\title{Custom classification metrics for h2o automl models}
\usage{
custom_h2o_confusion_metrics(
  model = NULL,
  newdata = NULL,
  threshold = c(0.4, 0.6)
)
}
\arguments{
\item{model}{An existin h2o model object.}

\item{newdata}{An h2o data object to apply the model object to. Typically a test set.}

\item{threshold}{A range of threshold/cut-off values to evaluate.}
}
\value{
A data.frame containing metrics for each threshold value provided.
}
\description{
model: The model name
threshold: Threshold value
logloss: The logloss of the model (see \url{https://www.kaggle.com/dansbecker/what-is-log-loss} for more information)
precision: True Positives / Predicted Positives
recall: True Positives / All Positives
neg_precision: Negative Prevision, True Negatives / Predicted Negatives
specificity: True Negatives / All Negatives
fall_out: 1 - \code{specificity}, used to plot ROC curves
f1_score: F1 Score or Balanced Accuracy, the harmonic mean of \code{precision} and \code{recall}
true_positives: Number of True Positives
false_negatives: Number of False Negatives
false_positives: Number of False Positives
true_negatives: Number of True Negatives
}
\examples{
\dontrun{custom_h2o_confusion_metrics(model = NA, newdata = NA, threshold = c(0,1)}
}
