% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_h2o_models_varimport.R
\name{save_h2o_models_varimport}
\alias{save_h2o_models_varimport}
\title{Save h2o automl models and variable importance}
\usage{
save_h2o_models_varimport(
  leaderboard,
  max_top_models = NA,
  path_slug,
  time_stamp = TRUE
)
}
\arguments{
\item{leaderboard}{A leaderboard object generated from \code{h2o::automl}.}

\item{max_top_models}{Number of models to save and generate variable importance for. If NA, all
models will be processed.}

\item{path_slug}{A file path where the models should be saved.}

\item{time_stamp}{If TRUE, a time stamp is constructed using \code{Sys.time}. User can also
provide their own time stamp as a string e.g. '2020_08_01'.}
}
\value{
Function will return a list with an element for each model containing the model object
and the variable importance of the model (if appropriate--not availabel for Ensemble and GLM). Individual
models will be saved in the directory provided by the \code{path_slug} argument, and a list with all model
objects and variable importane will also be saved to the \code{path_slug} directory.
}
\description{
Save h2o automl models and variable importance
}
\examples{
\dontrun{save_h2o_models_varimport(leaderboard, max_top_models = NA,
path_slug, time_stamp = TRUE)}
}
